requirements.txt
PyQt5
matplotlib
pandas
python-dateutil
numpy
scipy
matplotlib
fbs
SQLAlchemy
Flask-MySQLdb
requests
pyqtdeploy
bbfreeze
qrcode
pillow
pyqrcode
pypng
PyQt-builder